+++ packages/react-core/src/core/hooks/wallet-hooks.ts
@@ -13,6 +13,7 @@ import type {
   OKXWallet,
   PaperWallet,
   PhantomWallet,
+  RabbyWallet,
   RainbowWallet,
   SafeWallet,
   SmartWallet,
@@ -45,6 +46,7 @@ type WalletIdToWalletTypeMap = {
   phantom: PhantomWallet;
   walletConnectV1: WalletConnect;
   okx: OKXWallet;
+  rabby: RabbyWallet;
 };
 
 /**