+++ packages/react/src/wallet/wallets/smartWallet/SmartWalletConnecting.tsx
@@ -4,9 +4,13 @@ import { Button } from "../../../components/buttons";
 import { iconSize, spacing, fontSize } from "../../../design-system";
 import { ExclamationTriangleIcon } from "@radix-ui/react-icons";
 import {
+  useChain,
   useConnect,
   useConnectionStatus,
+  useNetworkMismatch,
   useWalletContext,
+  useWallet,
+  useSwitchChain,
   WalletConfig,
 } from "@thirdweb-dev/react-core";
 import { useCallback, useContext, useEffect, useRef, useState } from "react";
@@ -20,32 +24,31 @@ export const SmartWalletConnecting: React.FC<{
   onBack: () => void;
   onConnect: () => void;
   smartWallet: SmartWalletConfig;
-  personalWalletConfig: WalletConfig;
+  personalWallet: WalletConfig;
 }> = (props) => {
   const locale = useTWLocale().wallets.smartWallet;
+  const activeWallet = useWallet(); // personal wallet
 
-  // personal wallet
-  const { personalWalletInfo } = useWalletContext();
-  const personalWallet = personalWalletInfo?.wallet;
-  const personalWalletChainId = personalWalletInfo?.chainId;
+  const connect = useConnect();
+  const connectedChain = useChain();
+  const targetChain = useWalletContext().activeChain;
+
+  const mismatch = useNetworkMismatch();
+
+  const [connectError, setConnectError] = useState(false);
   const [switchError, setSwitchError] = useState(false);
-  const switchPersonalWalletChain = personalWalletInfo.switchChain;
   const [switchingNetwork, setSwitchingNetwork] = useState(false);
 
-  // smart wallet
-  const connect = useConnect();
-  const targetChain = useWalletContext().activeChain;
-  const mismatch = targetChain.chainId !== personalWalletChainId;
-  const [connectError, setConnectError] = useState(false);
   const connectionStatus = useConnectionStatus();
 
   const { onConnect } = props;
   const connectStarted = useRef(false);
 
+  const switchChain = useSwitchChain();
   const modalSize = useContext(ModalConfigCtx).modalSize;
 
   const handleConnect = useCallback(async () => {
-    if (!personalWallet || !personalWalletChainId || connectStarted.current) {
+    if (!activeWallet || !connectedChain || connectStarted.current) {
       return;
     }
     setConnectError(false);
@@ -53,26 +56,20 @@ export const SmartWalletConnecting: React.FC<{
     try {
       connectStarted.current = true;
       await connect(props.smartWallet, {
-        personalWallet: personalWallet,
+        personalWallet: activeWallet,
       });
       onConnect();
     } catch (e) {
       console.error(e);
       setConnectError(true);
     }
-  }, [
-    personalWallet,
-    personalWalletChainId,
-    connect,
-    props.smartWallet,
-    onConnect,
-  ]);
+  }, [activeWallet, connectedChain, connect, props.smartWallet, onConnect]);
 
   useEffect(() => {
     if (!mismatch) {
       handleConnect();
     }
-  }, [mismatch, handleConnect, personalWallet, personalWalletChainId]);
+  }, [mismatch, handleConnect, activeWallet, connectedChain]);
 
   if (!connectError && (connectionStatus === "connecting" || !mismatch)) {
     return (
@@ -114,8 +111,8 @@ export const SmartWalletConnecting: React.FC<{
     <Container fullHeight animate="fadein" flex="column">
       <Container p="lg">
         <ModalHeader
-          title={props.personalWalletConfig.meta.name}
-          imgSrc={props.personalWalletConfig.meta.iconURL}
+          title={props.personalWallet.meta.name}
+          imgSrc={props.personalWallet.meta.iconURL}
           onBack={props.onBack}
         />
       </Container>
@@ -153,14 +150,14 @@ export const SmartWalletConnecting: React.FC<{
                 gap: spacing.sm,
               }}
               onClick={async () => {
-                if (!personalWallet) {
+                if (!activeWallet) {
                   throw new Error("No active wallet");
                 }
                 setConnectError(false);
                 setSwitchError(false);
                 setSwitchingNetwork(true);
                 try {
-                  await switchPersonalWalletChain(targetChain.chainId);
+                  await switchChain(targetChain.chainId);
                 } catch (e) {
                   setSwitchError(true);
                 } finally {